+++ sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/client/ObservabilityContext.kt
@@ -4,6 +4,7 @@ import android.app.Application
 import com.launchdarkly.logging.LDLogger
 import com.launchdarkly.observability.api.ObservabilityOptions
 import io.opentelemetry.android.session.SessionManager
+import io.opentelemetry.api.common.Attributes
 
 /**
  * Shared information between plugins.
@@ -14,4 +15,5 @@ data class ObservabilityContext(
     val application: Application,
     val logger: LDLogger,
     var sessionManager: SessionManager? = null,
+    var resourceAttributes: Attributes = Attributes.empty(),
 )